timer and resend added to the otp login
