+++ package.json
@@ -29,6 +29,7 @@
     "react-otp-input": "^3.0.4",
     "react-redux": "^8.0.4",
     "react-select": "^5.4.0",
+    "react-timer-hook": "^3.0.7",
     "react-toastify": "^9.0.7",
     "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
