+++ styles/pages/auth/login-otp.module.scss
@@ -66,6 +66,13 @@
     margin-bottom: 10px;
   }
 
+  .resend{
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+    align-items: center;
+  }
+
 }
 
 .footer {
